requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
python-dateutil==2.8.2
